requirements.txt
pygame-ce>=2.5.0
numpy<2.0.0
Pillow==10.1.0
PyOpenGL==3.1.7
PyOpenGL_accelerate==3.1.7